fix(tests): resolve ~1200 test tsc errors — JSDoc→TS casts across 259 files
Agents converted JSDoc type casts to TypeScript across the test suite:
- `/** @type {any} */ (expr)` → `(expr as any)`
- `/** @type {*} */` → `as any`
- JSDoc param types → TS param annotations
- TS2345/TS2322/TS2339 fixes with proper casts

Fixed seekDiff.test.ts syntax error from agent codemod.

Source: 0 tsc. Test: 908 tsc (down from 2,086). Tests: 356/356.

+++ test/unit/domain/services/BitmapIndexBuilder.integrity.test.ts
@@ -39,7 +39,7 @@ describe('BitmapIndexBuilder Integrity Tests', () => {
       expect(Object.keys(tree1).sort()).toEqual(Object.keys(tree2).sort());
 
       for (const key of Object.keys(tree1)) {
-        expect(/** @type {Uint8Array} */ (tree1[key]).toString()).toBe(/** @type {Uint8Array} */ (tree2[key]).toString());
+        expect((tree1[key] as Uint8Array).toString()).toBe((tree2[key] as Uint8Array).toString());
       }
     });
